Modify onClickConnectTest
