+++ app/src/androidTest/java/com/github/dedis/student20_pop/MainActivityTest.java
@@ -1,8 +1,10 @@
 package com.github.dedis.student20_pop;
 
 import android.Manifest;
+import android.content.pm.PackageManager;
 import android.view.View;
 
+import androidx.core.content.ContextCompat;
 import androidx.test.core.app.ActivityScenario;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.rule.GrantPermissionRule;
@@ -25,6 +27,7 @@ import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.not;
 
 public class MainActivityTest {
+
     @Rule
     public final GrantPermissionRule rule = GrantPermissionRule.grant(Manifest.permission.CAMERA);
     @Rule
@@ -57,7 +60,13 @@ public class MainActivityTest {
     @Test
     public void onClickConnectTest() {
         onView(withId(R.id.tab_connect)).perform(click());
-        onView(withId(R.id.fragment_qrcode)).check(matches(isDisplayed()));
+        if (ContextCompat.checkSelfPermission(PoPApplication.getAppContext(),
+                Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
+            onView(withId(R.id.fragment_qrcode)).check(matches(isDisplayed()));
+        }
+        else {
+            onView(withId(R.id.fragment_camera_perm)).check(matches(isDisplayed()));
+        }
     }
 
     @Test
