+++ app/src/androidTest/java/com/github/dedis/student20_pop/ui/event/creation/MeetingEventCreationFragmentTest.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui;
+package com.github.dedis.student20_pop.ui.event.creation;
 
 import android.view.View;
 import android.widget.DatePicker;
@@ -38,7 +38,7 @@ import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.isIn;
 import static org.hamcrest.Matchers.not;
 
-public class RollCallEventCreationFragmentTest {
+public class MeetingEventCreationFragmentTest {
     private final int YEAR = 2022;
     private final int MONTH_OF_YEAR = 10;
     private final int DAY_OF_MONTH = 10;
@@ -46,7 +46,6 @@ public class RollCallEventCreationFragmentTest {
     private final int HOURS = 12;
     private final int MINUTES = 15;
     private final String TIME = "" + HOURS + ":" + MINUTES;
-
     @Rule
     public ActivityScenarioRule<OrganizerActivity> activityScenarioRule =
             new ActivityScenarioRule<>(OrganizerActivity.class);
@@ -68,28 +67,28 @@ public class RollCallEventCreationFragmentTest {
         });
 
         onView(allOf(withId(R.id.add_future_event_button), withEffectiveVisibility(ViewMatchers.Visibility.VISIBLE))).perform(click());
-        onView(withText(getApplicationContext().getString(R.string.roll_call_event))).perform(click());
+        onView(withText(getApplicationContext().getString(R.string.meeting_event))).perform(click());
     }
 
     @Test
     public void canLaunchEventMeetingFragment() {
-        onView(withId(R.id.fragment_create_roll_call_event)).check(matches(isDisplayed()));
+        onView(withId(R.id.fragment_meeting_event_creation)).check(matches(isDisplayed()));
     }
 
     @Test
     public void confirmButtonIsDisabled() {
-        onView(withId(R.id.roll_call_confirm)).check(matches(not(isEnabled())));
+        onView(withId(R.id.meeting_event_creation_confirm)).check(matches(not(isEnabled())));
     }
 
     @Test
     public void cancelButtonWorks() {
-        onView(withId(R.id.roll_call_cancel)).perform(click());
+        onView(withId(R.id.meeting_event_creation_cancel)).perform(click());
         onView(withId(R.id.fragment_organizer)).check(matches(isDisplayed()));
     }
 
     @Test
     public void confirmButtonIsEnabledWhenRequiredFieldsFilled() {
-        onView(withId(R.id.roll_call_title_text)).perform(typeText("Random meeting title"));
+        onView(withId(R.id.meeting_title_text)).perform(typeText("Random meeting title"));
 
         onView(withId(R.id.start_date_edit_text)).perform(click());
         onView(withClassName(Matchers.equalTo(DatePicker.class.getName()))).perform(PickerActions.setDate(YEAR, MONTH_OF_YEAR, DAY_OF_MONTH));
@@ -101,13 +100,13 @@ public class RollCallEventCreationFragmentTest {
         onView(withId(android.R.id.button1)).perform(click());
         onView(withId(R.id.start_time_edit_text)).check(matches(withText(TIME)));
 
-        onView(withId(R.id.roll_call_confirm)).check(matches(isEnabled()));
+        onView(withId(R.id.meeting_event_creation_confirm)).check(matches(isEnabled()));
     }
 
     @Test
     public void confirmAddsEventToEventList() {
-        final String RANDOM_EVENT_TITLE = "Random roll call title";
-        onView(withId(R.id.roll_call_title_text)).perform(typeText(RANDOM_EVENT_TITLE));
+        final String RANDOM_EVENT_TITLE = "Random meeting title";
+        onView(withId(R.id.meeting_title_text)).perform(typeText(RANDOM_EVENT_TITLE));
 
         onView(withId(R.id.start_date_edit_text)).perform(click());
         onView(withClassName(Matchers.equalTo(DatePicker.class.getName()))).perform(PickerActions.setDate(YEAR, MONTH_OF_YEAR, DAY_OF_MONTH));
@@ -119,8 +118,8 @@ public class RollCallEventCreationFragmentTest {
         onView(withId(android.R.id.button1)).perform(click());
         onView(withId(R.id.start_time_edit_text)).check(matches(withText(TIME)));
 
-        onView(withId(R.id.roll_call_confirm)).check(matches(isEnabled()));
-        onView(withId(R.id.roll_call_confirm)).perform(click());
+        onView(withId(R.id.meeting_event_creation_confirm)).check(matches(isEnabled()));
+        onView(withId(R.id.meeting_event_creation_confirm)).perform(click());
 
         activityScenarioRule.getScenario().onActivity(
                 activity -> {
