+++ app/src/androidTest/java/com/github/dedis/student20_pop/ui/event/creation/PollEventCreationFragmentTest.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui;
+package com.github.dedis.student20_pop.ui.event.creation;
 
 import android.view.View;
