+++ app/src/androidTest/java/com/github/dedis/student20_pop/ui/qrcode/QRCodeScanningFragmentTest.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui;
+package com.github.dedis.student20_pop.ui.qrcode;
 
 import android.Manifest;
